Move open_meteo to a directory that makes sense
